Firefox updater full auto + more checks
Curl pipes the Firefox download page (I know there’s an xml file I’m
missing somewhere) to grep until the first match that looks like a
version number. Throws it in the URL. Also checks to see if the current
version matches the version on the machine incase download isn’t
necessary.

+++ autoUpdateFirefox/autoUpdateFirefox.sh
@@ -1,20 +1,23 @@
 #!/bin/sh
 # Firefox Auto-Install/Update
 
-echo 'Dowloading new version of Firefox...'
-firefoxVersion='29.0.1'
-# download firefox
-curl -GLo /tmp/firefox.dmg "https://download.mozilla.org/?product=firefox-$firefoxVersion-SSL&os=osx&lang=en-US"
-# mount image
-hdiutil mount -quiet -nobrowse /tmp/firefox.dmg
-# delete old version if it exists
-if [[ $1 == '2' ]]; then
-	echo 'Uninstalling current version...'
-	sudo rm -rf /Applications/Firefox.app
-fi
-echo 'Installing up-to-date version....'
-sudo cp -R /Volumes/Firefox/Firefox.app /Applications
-hdiutil unmount /Volumes/Firefox
-rm /tmp/firefox.dmg
+currentFirefoxVersion=`curl -silent http://www.mozilla.org/en-US/firefox/all/ | grep -m1 -o '[0-9]*\.[0-9]\.[0-9]'`
+localFirefoxVersion=`defaults read /Applications/Firefox.app/Contents/Info.plist CFBundleShortVersionString`
+echo "Installing/Updating Firefox to version $currentFirefoxVersion"
 
+if [[ "$currentFirefoxVersion" == "$localFirefoxVersion" ]]; then
+	echo "Firefox is already up-to-date! Exiting..."
+	exit 1
+else
+	if [[ -d /Applications/Firefox.app/ ]]; then
+		echo 'Removing old version...'
+		sudo rm -rf /Applications/Firefox.app
+	fi
+	curl -GLo /tmp/firefox.dmg "https://download.mozilla.org/?product=firefox-$currentFirefoxVersion-SSL&os=osx&lang=en-US"
+	hdiutil mount -quiet -nobrowse /tmp/firefox.dmg
+	sudo cp -R /Volumes/Firefox/Firefox.app /Applications
+	hdiutil unmount /Volumes/Firefox
+	rm /tmp/firefox.dmg
+	echo "Firefox updated! Exiting..."
+fi
 exit 0
